requirements.txt
Django>=3.2.4,<3.3
djangorestframework>=3.12.4,<3.13
django-filter>=22.1,<22.2
drf-spectacular>=0.15.1,<0.16
graphene-django==2.13.0